componentizando o loading
